cemini_mcp/requirements.txt
fastmcp==3.1.0
redis>=4.6.0
psycopg2-binary>=2.9.0
